+++ system/application/controllers/Event.php
@@ -3,6 +3,9 @@ class Event extends Controller {
 
 	function show_all()
 	{
+	    if ($this->session->userdata('admin') == FALSE)
+	        redirect('dashboard/index');
+	        
 		$this->load->model('Event_model');
 		$data['result'] = $this->Event_model->get_all_events();
 		$this->load->view('header');
@@ -28,6 +31,9 @@ class Event extends Controller {
 
 	function insert()
 	{
+	    if ($this->session->userdata('admin') == FALSE)
+	        redirect('dashboard/index');
+	        
 		$this->load->view('header');
 		$this->load->view('Event/Event_insert_form');
 		$this->load->view('footer');
@@ -35,6 +41,9 @@ class Event extends Controller {
 	
 	function create()
 	{
+	    if ($this->session->userdata('admin') == FALSE)
+	        redirect('dashboard/index');
+	    
 		$this->load->model('Event_model');
         
 		$name = $this->input->post('name');
@@ -55,6 +64,9 @@ class Event extends Controller {
 	
 	function delete($id)
 	{
+	    if ($this->session->userdata('admin') == FALSE)
+	        redirect('dashboard/index');
+	        
 		$this->load->model('Event_model');
 		$this->Event_model->delete_event($id);
 		$data['result'] = $this->Event_model->get_all_events();
